Timeslot race condition fix (#405)
* Time slot fix
* Commenting out logs
* Adding appointment conflict check on appointment POST and PUT
